# apps/docs/docs/api.yaml
openapi: 3.0.1
info:
  title: Chatvolt AI - API Specifications
  description: |-

  termsOfService: http://chatvolt.ai/terms
  contact:
    email: support@chatvolt.ai
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://api.chatvolt.ai
...
paths:
  /datasources:
    post:
      tags:
        - datasources
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                fileName:
                  type: string
                type:
                  enum:
                    - file
                datastoreId:
                  type: string
                custom_id:
                  type: string
                  description: Useful for multi-tenant setups

              required:
                - type
                - datastoreId
                - file
          application/json:
            schema:
              type: object
              properties:
                type:
                  enum: ['web_page', 'web_site']
                datastoreId:
                  type: string
                custom_id:
                  type: string
                  description: Useful for multi-tenant setups
                config:
                  type: object
                  properties:
                    source_url:
                      type: string
                      format: url
                      description: web_page, web_site
                    sitemap:
                      type: string
                      format: url
                      description: web_site

              # oneOf:
              #   - $ref: '#/components/schemas/DatasourceWebPage'
              #   - $ref: '#/components/schemas/DatasourceWebSite'
              # discriminator:
              #   propertyName: type
              #   mappping:
              #     "web_page": '#/components/schemas/y'
              #     "web_site": '#/components/schemas/DatasourceWebSite'

              required:
                - type
                - datastoreId
                - config
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datasource'
        '400':
          description: Invalid body
        '403':
          description: Unauthorized
  /datasources/{id}:
    get:
      tags:
        - datasources
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: ID of the datasource
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datasource'
        '403':
          description: Unauthorized
    delete:
      tags:
        - datasources
      parameters:
        - in: path
          name: id
          schema:
            type: string
            description: ID of the datasource
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Datasource'
        '403':
          description: Unauthorized
...
components:
  schemas:
    Datasource:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
        status:
          enum:
            [unsynched, pending, running, synched, error, usage_limit_reached]
        groupId:
          type: string
        updatedAt:
          type: string
          format: date
        createdAt:
          type: string
          format: date
        lastSynch:
          type: string
          format: date
        config:
          type: object
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
security:
  - bearerAuth: []